Moved all preference classes into prefs package
