+++ app/src/main/AndroidManifest.xml
@@ -177,7 +177,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".ui.PreferencesActivity">
+        <activity android:name=".ui.prefs.PreferencesActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MANAGE_NETWORK_USAGE" />
                 <category android:name="android.intent.category.DEFAULT" />
