+++ app/src/main/java/org/kontalk/ui/prefs/RootPreferenceFragment.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.kontalk.ui;
+package org.kontalk.ui.prefs;
 
 import com.github.machinarius.preferencefragment.PreferenceFragment;
 
@@ -72,7 +72,7 @@ public class RootPreferenceFragment extends PreferenceFragment {
         return super.onOptionsItemSelected(item);
     }
 
-    public void invokeCallback(int key) {
+    void invokeCallback(int key) {
         if (mCallback != null)
             mCallback.onNestedPreferenceSelected(key);
     }
